[rename] #24 프로젝트 컨트롤러로 병합
