+++ src/main/java/org/umc/peerre/domain/project/controller/ProjectController.java
@@ -3,10 +3,13 @@ package org.umc.peerre.domain.project.controller;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.umc.peerre.domain.project.dto.request.CreateCommentRequestDto;
 import org.umc.peerre.domain.project.dto.request.CreateProjectRequestDto;
+import org.umc.peerre.domain.project.dto.response.CreateCommentResponseDto;
 import org.umc.peerre.domain.project.dto.response.CreateProjectResponseDto;
 import org.umc.peerre.domain.project.dto.response.MyFeedbackResponseDto;
 import org.umc.peerre.domain.project.dto.response.TeamInfoResponseDto;
+import org.umc.peerre.domain.project.service.CommentService;
 import org.umc.peerre.domain.project.service.ProjectService;
 import org.umc.peerre.global.common.SuccessResponse;
 import org.umc.peerre.global.config.auth.UserId;
@@ -19,11 +22,19 @@ import java.util.List;
 public class ProjectController {
 
     private final ProjectService projectService;
+    private final CommentService commentService;
+
 
     @PostMapping
     public ResponseEntity<SuccessResponse<?>> createProject(@RequestBody CreateProjectRequestDto createProjectRequestDto) {
         final CreateProjectResponseDto newProject = projectService.createProject(createProjectRequestDto);
         return SuccessResponse.created(newProject);
+    }
+
+    @PostMapping("/comments")
+    public ResponseEntity<SuccessResponse<?>> createComments(@UserId Long userId, @RequestBody CreateCommentRequestDto createCommentRequestDto) {
+        CreateCommentResponseDto newComment = commentService.createComment(userId,createCommentRequestDto);
+        return SuccessResponse.created(newComment);
     }
 
     @PostMapping("/{projectId}")
